+++ extensions/vscode/src/multiStepInputs/newDeployment.ts
@@ -61,9 +61,11 @@ import {
 import { ENTRYPOINT_FILE_EXTENSIONS } from "src/constants";
 import { extensionSettings } from "src/extension";
 import {
+  authConnectCloud,
   fetchSnowflakeConnections,
   findExistingCredentialByURL,
   isConnect,
+  isConnectCloud,
   isSnowflake,
   platformList,
 } from "src/multiStepInputs/common";
@@ -199,7 +201,10 @@ export async function newDeployment(
       credentialListItems = credentials.map((credential) => ({
         iconPath: new ThemeIcon("posit-publisher-posit-logo"),
         label: credential.name,
-        description: credential.url,
+        description:
+          credential.serverType === ServerType.CONNECT_CLOUD
+            ? `${credential.accountName} | Posit Connect Cloud`
+            : credential.url,
       }));
       credentialListItems.push({
         iconPath: new ThemeIcon("plus"),
@@ -583,6 +588,22 @@ export async function newDeployment(
       serverType = enumKey ? ServerType[enumKey] : ServerType.CONNECT;
       platformName = pick.label as PlatformName;
 
+      if (isConnectCloud(serverType)) {
+        try {
+          await showProgress(
+            "Authenticating with Connect Cloud",
+            viewId,
+            async () => await authConnectCloud(),
+          );
+          // TODO: fetch the accounts using the token
+        } catch {
+          // errors have already been displayed by authConnectCloud
+          return;
+        }
+        // bail out for now
+        return Promise.resolve(undefined);
+      }
+
       return (input: MultiStepInput) => inputServerUrl(input, state);
     }
 
@@ -943,6 +964,10 @@ export async function newDeployment(
         newDeploymentData.newCredentials.url,
         newDeploymentData.newCredentials.apiKey,
         newDeploymentData.newCredentials.snowflakeConnection,
+        "",
+        "",
+        "",
+        "",
         serverType,
       );
       newOrSelectedCredential = response.data;